First draft of new insights package
