+++ setup.cfg
@@ -4,7 +4,7 @@ style = pep440
 versionfile_source = asreviewcontrib/visualization/_version.py
 versionfile_build = asreviewcontrib/visualization/_version.py
 tag_prefix = v
-parentdir_prefix = asreview-visualization-
+parentdir_prefix = asreview-insights-
 
 [flake8]
 max-line-length = 88
