+++ tests/README.md
@@ -1,0 +1,8 @@
+# Testing ASReview-insights
+
+## Create ASReview files
+
+```
+# asreview simulate benchmark:van_de_schoot_2017 -s tests/asreview_files/sim_ptsd_2.asreview
+asreview simulate tests/ptsd.csv -s tests/asreview_files/sim_ptsd_2.asreview
+```
