+++ tests/asreview_files/sim_ptsd_1.asreview/project.json
@@ -1,0 +1,1 @@
+{"version": "1.0a0+7.g93d63aa", "id": "sim_ptsd_2", "mode": "simulate", "name": null, "description": null, "authors": null, "created_at_unix": 1640867659, "datetimeCreated": "2021-12-30 13:34:19.415889", "projectInitReady": false, "reviewFinished": false, "reviews": [{"id": "05406625aa954cb49308e2bc96637494", "start_time": "2021-12-30 13:34:19.610057", "review_finished": true, "end_time": "2021-12-31 17:15:13.487972"}], "feature_matrices": [{"id": "tfidf", "filename": "tfidf_feature_matrix.npz"}]}
