+++ tests/asreview_files/sim_ptsd_2.asreview/project.json
@@ -1,0 +1,1 @@
+{"version": "1.0a0+7.g93d63aa", "id": "sim_ptsd_2", "mode": "simulate", "name": null, "description": null, "authors": null, "created_at_unix": 1641287810, "datetimeCreated": "2022-01-04 10:16:50.243199", "projectInitReady": false, "reviewFinished": false, "reviews": [{"id": "433c49377a81408baa0d3af3ade4fa5b", "start_time": "2022-01-04 10:16:50.459762", "review_finished": true, "end_time": "2022-01-04 10:26:02.213440"}], "feature_matrices": [{"id": "tfidf", "filename": "tfidf_feature_matrix.npz"}]}
