+++ tests/asreview_files/sim_ptsd_2.asreview/reviews/433c49377a81408baa0d3af3ade4fa5b/settings_metadata.json
@@ -1,0 +1,1 @@
+{"settings": {"model": "nb", "query_strategy": "max", "balance_strategy": "double", "feature_extraction": "tfidf", "n_instances": 1, "n_prior_included": 1, "n_prior_excluded": 1, "mode": "simulate", "model_param": {"alpha": 3.822}, "query_param": {}, "feature_param": {"ngram_max": 1, "stop_words": "english", "split_ta": 0, "use_keywords": 0}, "balance_param": {"a": 2.155, "alpha": 0.94, "b": 0.789, "beta": 1.0}}, "state_version": "1", "software_version": "1.0a0+7.g93d63aa"}
